% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dac_odbc.R
\name{dac_cfg_connect}
\alias{dac_cfg_connect}
\title{Connect to the DHSC analytical cloud data store via ODBC driver using a config yaml file}
\usage{
dac_cfg_connect(
  config_yml,
  env = c("prod", "qa", "test", "dev"),
  use_cache = TRUE,
  connections_pane = FALSE
)
}
\arguments{
\item{config_yml}{yaml file with configuration settings.}

\item{env}{development environment to use. One of "prod" (default), "qa",
"test", or "dev".}

\item{use_cache}{should any existing cached tokens be used for
authentication.}

\item{connections_pane}{whether to add the connection to RStudio's
interactive connection pane (adds additional loading time).}
}
\value{
a \code{DBIConnection} object equivalent to that returned by
\code{\link[DBI]{dbConnect}}
}
\description{
Connect to the DHSC analytical cloud data store via ODBC driver using a config yaml file
}
