% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dac_connect.R
\name{dac_connect}
\alias{dac_connect}
\title{Connect to the DHSC analytical cloud data store via ODBC driver}
\usage{
dac_connect(use_cache = TRUE, connections_pane = FALSE)
}
\arguments{
\item{use_cache}{Should any existing cached tokens be used for
authentication.}

\item{connections_pane}{Whether to add the connection to RStudio's
interactive connection pane (adds additional loading time).}
}
\value{
a \code{DBIConnection} object equivalent to that returned by
\code{DBI::dbConnect}
}
\description{
Connects to the Databricks SQL endpoint on the DHSC analytical
cloud (DAC). Connections are authenticated using Azure Entra ID.
}
\details{
Please ensure that you have the correct .Renviron variables
set.IMPORTANT - NEVER COMMIT YOUR .Renviron FILE
}
