% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{get_proxy_config}
\alias{get_proxy_config}
\title{Automatically get proxy config}
\usage{
get_proxy_config(target_url = "https://www.google.com")
}
\arguments{
\item{target_url}{url with host for which to lookup the proxy server}
}
\value{
curl config parameters
}
\description{
Automatically get proxy config
}
