% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_vault.R
\name{key_vault}
\alias{key_vault}
\title{Get DAC Azure Key Vault client}
\usage{
key_vault(...)
}
\arguments{
\item{...}{Further arguments passed to \code{AzureKeyVault::key_vault}}
}
\value{
Key Vault client
}
\description{
Note that retrieving client and any secrets will only work
it the correct proxy is set - see \code{DHSCdatatools::get_proxy_config}
for getting the correct proxy configuration.
}
