% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{set_proxy_config}
\alias{set_proxy_config}
\title{Correctly set the DHSC proxy for base R methods}
\usage{
set_proxy_config(target_url = "https://www.google.com")
}
\arguments{
\item{target_url}{url with host for which to lookup the proxy server}
}
\value{
invisibility, the old global config.
}
\description{
Configure proxy settings for functions based on curl using Windows settings.
}
\details{
While R uses the \code{wininet} download method that correctly obtains
proxy details from Windows via the WinINet functions, other packages
such as \pkg{httr} are based on \pkg{curl} and so need the proxy settings
configured to use the corporate proxy and firewall. This function uses
\code{\link[curl]{ie_get_proxy_for_url}} to read these settings and then applies
them using \code{\link[httr]{set_config}}.
}
