% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyring.R
\name{set_token}
\alias{set_token}
\title{Helper function to save DAC Databrick tokens to keyring}
\usage{
set_token(
  env = c("prod", "qa", "test", "dev"),
  username = NULL,
  keyring = NULL
)
}
\arguments{
\item{env}{development environment of token. One of "prod" (default), "qa",
"test", or "dev".}

\item{username}{username, a character scalar, or NULL if the key is not
associated with a username passed to \code{\link[keyring]{key_set}}.}

\item{keyring}{for systems that support multiple keyrings, specify the name
of the keyring to use here. If NULL, then the default keyring is used.
Passed to \code{\link[keyring]{key_set}}.}
}
\description{
Helper function to save DAC Databrick tokens to keyring
}
