% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparklyr.R
\name{sparklyr_cfg_connect}
\alias{sparklyr_cfg_connect}
\title{Initialise sparklyr and open connection using a config yaml file}
\usage{
sparklyr_cfg_connect(
  config_yml,
  env = c("prod", "qa", "test", "dev"),
  use_cache = TRUE
)
}
\arguments{
\item{config_yml}{yaml file with configuration settings.}

\item{env}{development environment to use from file. One of "prod" (default),
"qa", "test", or "dev".}

\item{use_cache}{should any existing cached Azure tokens be used for
authentication when retrieving config data from Key Vault.}
}
\value{
a \pkg{sparklyr} connection object.
}
\description{
Initialise sparklyr and open connection using a config yaml file
}
