% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparklyr.R
\name{sparklyr_connect}
\alias{sparklyr_connect}
\title{Initialise sparklyr and open connection}
\usage{
sparklyr_connect(
  env = c("prod", "qa", "test", "dev"),
  token = NULL,
  use_cache = TRUE,
  reset_conda = FALSE,
  install_ml = FALSE
)
}
\arguments{
\item{env}{development environment to use. One of "prod" (default), "qa",
"test", or "dev".}

\item{token}{Databricks developer token. If not provided windows
credential manager (keyring) is used see \link{set_token} for setting tokens.}

\item{use_cache}{should any existing cached Azure tokens be used for
authentication when retrieving config data from Key Vault.}

\item{reset_conda}{should existing conda environment be deleted first.}

\item{install_ml}{installs ML related Python libraries.}
}
\value{
a \pkg{sparklyr} connection object.
}
\description{
Function will download relevant configuration from the DAC Key Vault and use
this to create a Python conda environment in which to run the Databricks
connection.
}
