% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSCdata-package.R
\docType{package}
\name{DHSCdatatools-package}
\alias{DHSCdatatools}
\alias{DHSCdatatools-package}
\title{DHSCdatatools: Collection of tools for working with the DHSC analytical cloud (DAC)}
\description{
This package will be a place to collect various tools that can be used with the DHSC analytical cloud (DAC) platform. Currently, it includes tools for connecting to the DAC SQL endpoint and running sparklyr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DataS-DHSC/DHSCdatatools}
  \item Report bugs at \url{https://github.com/DataS-DHSC/DHSCdatatools/issues}
}

}
\author{
\strong{Maintainer}:  Data Science Hub \email{datascience@dhsc.gov.uk}

}
\keyword{internal}
